requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==4.0.3
django-csp==3.7
django-extensions==3.1.5
googlemaps==4.6.0
idna==3.3
joblib==1.1.0
nltk==3.7
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
urllib3==1.26.9
wikipedia==1.4.0
